Make App Material
Added Theme color, this invloves textColors, accent color, etc
Changed the color of the textview in cardView. one was for primary text, the other was for secondary text
Changed the toolbar color to color_primary of the theme
Added Constraint Layout to keep the toolbar and swipeRefresh in order
Made the swipeRefresh to work properly by adding eventListener to refresh the view by calling the service that makes a asyncTask, returns the books from the internet server, deletes the existing data in the localDB, populates the localDB with the new data fetched.