Make App Material (Home page)
Added color theme to the AppTheme in style.xml, this involves colorPrimary, colorPrimaryDark, accent color, etc. Color theme does not impact usability of the app.
Added the AppBar which is associated with toolbar. Added LinearLayout with vertical orientation to make the AppBar and Scrollview align(removed the marginTop which has the length of the appbar previously).
Added scaleType to Image in the CardView(list_item) to fit the Card properly(full bleed). Made the image width(wrap_content) to be the width of the card.