Fixed Performance issues in detail page (When you swipe between articles the app freezes)
Splited the huge text body into multiple smaller chunks, in a BACKGROUND thread), place a RecyclerView, set an adapter and pass an array of paragraphs to it.

FYI, i had to remake the detail page, implemented CoordinatorLayout + CollapsingToolbarLayout + RecyclerView, placed image, title and subtitle in the collapsing toolbar, and use the RecyclerView only to display the body text. But i and yet to make the 'parallax effect' where the title get pinned in the tool bar when collapsing
**Also fixed a progress bar showing when you swipe to the first and last article.