Implemented the Article title to pin into the toolbar
Added a custom layout added to the CollapsingToolbar and pin it. I had to delete irrelevant style files like for version 19 and 21

Some of the notes to take care:
The the toolbar should have the height of your collapsed content, this can be done by code or at the XML.
The elements you want to hide must be BEFORE the toolbar definition in the layout XML file
The elements you want to remain visible must be AFTER the toolbar definition in the layout XML file.